+++ lib/src/http_result.dart
@@ -2,10 +2,10 @@ import 'package:http/http.dart';
 
 class HttpResult<T> {
   final Response response;
-  final T value;
+  final T data;
   final dynamic exception;
 
-  HttpResult({this.response, this.value, this.exception});
+  HttpResult({this.response, this.data, this.exception});
 
   bool get isSuccessful =>
       (response != null) &&
@@ -13,6 +13,13 @@ class HttpResult<T> {
       (response.statusCode <= 299);
 
   bool get isUnsuccessful => (response != null) && (response.statusCode >= 400);
+
+  String get body => (response != null) ? response.body : '';
+
+  Map<String, String> get headers =>
+      (response != null) ? response.headers : Map();
+
+  bool status(int code) => (response != null) && (response.statusCode == code);
 
   bool get hasFailed => (exception != null);
 }